src:iptables: drop /sbin -> /usr/sbin compat symlinks
Debian Buster included them, it should be enough to have a Debian release as
transition period. So drop them, no longer required.

The code in prerm is left so we can clean old symlinks when deleting the
package.

While at it, introduce a new NEWS entry to explain this and other important
changes in this release.

Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>