libiptc-dev: reintroduce package with shared pkg-config file
We can't get rid of the libiptc-dev package because pkg-config searchs for the
libiptc.pc file in some packages (like systemd).
The libiptc.pc file is very simple in the sense that it just points to libip4tc
and libip6tc.

My personal opinion is that the change should be made in upstream systemd
instead. They should stop depending on libiptc and directly use either libip4tc
or libip6tc, but that change is apparently too hard. So be it, let the package
contain just a single file.

Closes: #947176
Closes: #946202
Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>